+++ crates/miden-objects/src/account/storage/map.rs
@@ -1,4 +1,4 @@
-use miden_crypto::merkle::EmptySubtreeRoots;
+use miden_crypto::merkle::{EmptySubtreeRoots, MerkleError};
 
 use super::{
     ByteReader, ByteWriter, Deserializable, DeserializationError, Digest, Serializable, Word,
@@ -46,14 +46,15 @@ impl StorageMap {
         StorageMap { map: Smt::new() }
     }
 
-    pub fn with_entries(entries: impl IntoIterator<Item = (RpoDigest, Word)>) -> Self {
-        let mut storage_map = Smt::new();
-
-        for (key, value) in entries {
-            storage_map.insert(key, value);
-        }
-
-        StorageMap { map: storage_map }
+    /// Creates a new [`StorageMap`] from the provided key-value entries.
+    ///
+    /// # Errors
+    ///
+    /// Returns an error if the provided entries contain multiple values for the same key.
+    pub fn with_entries(
+        entries: impl IntoIterator<Item = (RpoDigest, Word)>,
+    ) -> Result<Self, MerkleError> {
+        Smt::with_entries(entries).map(|map| StorageMap { map })
     }
 
     // PUBLIC ACCESSORS
@@ -138,7 +139,8 @@ impl Deserializable for StorageMap {
 
 #[cfg(test)]
 mod tests {
-    use miden_crypto::{hash::rpo::RpoDigest, Felt};
+    use assert_matches::assert_matches;
+    use miden_crypto::{hash::rpo::RpoDigest, merkle::MerkleError, Felt};
 
     use super::{Deserializable, Serializable, StorageMap, Word, EMPTY_STORAGE_MAP_ROOT};
 
@@ -160,7 +162,7 @@ mod tests {
                 [Felt::new(5_u64), Felt::new(6_u64), Felt::new(7_u64), Felt::new(8_u64)],
             ),
         ];
-        let storage_map = StorageMap::with_entries(storage_map_leaves_2);
+        let storage_map = StorageMap::with_entries(storage_map_leaves_2).unwrap();
 
         let bytes = storage_map.to_bytes();
         assert_eq!(storage_map, StorageMap::read_from_bytes(&bytes).unwrap());
@@ -170,5 +172,23 @@ mod tests {
     fn test_empty_storage_map_constants() {
         // If these values don't match, update the constants.
         assert_eq!(StorageMap::default().root(), EMPTY_STORAGE_MAP_ROOT);
+    }
+
+    #[test]
+    fn account_storage_map_fails_on_duplicate_entries() {
+        // StorageMap with values
+        let storage_map_leaves_2: [(RpoDigest, Word); 2] = [
+            (
+                RpoDigest::new([Felt::new(101), Felt::new(102), Felt::new(103), Felt::new(104)]),
+                [Felt::new(1_u64), Felt::new(2_u64), Felt::new(3_u64), Felt::new(4_u64)],
+            ),
+            (
+                RpoDigest::new([Felt::new(101), Felt::new(102), Felt::new(103), Felt::new(104)]),
+                [Felt::new(5_u64), Felt::new(6_u64), Felt::new(7_u64), Felt::new(8_u64)],
+            ),
+        ];
+
+        let error = StorageMap::with_entries(storage_map_leaves_2).unwrap_err();
+        assert_matches!(error, MerkleError::DuplicateValuesForIndex(_));
     }
 }